take pic send to faceapiserver and it will return faceDescriptors
there is a bug that if the request to faceapiserver takes more than 30 secs, it will fail. this is because heroku will timeout server requests if they take more than 30 secs.

The length of time that the request takes is affected by how quick faceapi can process the input image, detect face and return the face descriptors.